requirements.txt
flask
openai
python-dotenv
apscheduler
google-genai